refaactor memberItem and departmentItem for own files
